+++ src/datamatrix.jl
@@ -1,21 +1,3 @@
-function _detect_var_id_cols(var)
-	id_cols = String[]
-	push!(id_cols, hasproperty(var, "id") ? "id" : first(names(var)))
-	hasproperty(var, :feature_type) && push!(id_cols, "feature_type")
-	@assert allunique(id_cols)
-
-	v = select(var, id_cols; copycols=false)
-	@assert size(unique(v),1) == size(v,1) "Failed to autodetect unique variable IDs (tried $id_cols)."
-	id_cols
-end
-
-function _detect_obs_id_cols(obs)
-	id_col = hasproperty(obs, "id") ? "id" : first(names(obs))
-	@assert allunique(obs[!,id_col]) "Failed to autodetect unique observation IDs (tried \"$id_col\")."
-	String[id_col]
-end
-
-
 """
 	struct DataMatrix{T,Tv,To}
 
@@ -25,8 +7,6 @@ Fields:
 * `matrix::T` - The matrix.
 * `var::Tv` - Variable annotations.
 * `obs::To` - Observation annotations.
-* `var_id_cols::Vector{String}` - Which column(s) to use as IDs.
-* `obs_id_cols::Vector{String}` - Which column(s) to use as IDs.
 * `models::Vector{ProjectionModel}` - Models used in the creation of this `DataMatrix`.
 
 The rows of the `var` and `obs` tables must be unique, considering only the `var_id_cols`/`obs_id_cols`.
@@ -35,10 +15,8 @@ struct DataMatrix{T,Tv,To}
 	matrix::T
 	var::Tv
 	obs::To
-	var_id_cols::Vector{String}
-	obs_id_cols::Vector{String}
 	models::Vector{ProjectionModel}
-	function DataMatrix(matrix::T, var::Tv, obs::To, var_id_cols, obs_id_cols, models) where {T,Tv,To}
+	function DataMatrix(matrix::T, var::Tv, obs::To, models) where {T,Tv,To}
 		P,N = size(matrix)
 		p = size(var,1)
 		n = size(obs,1)
@@ -46,29 +24,22 @@ struct DataMatrix{T,Tv,To}
 			throw(DimensionMismatch(string("Matrix has dimensions (",P,',',N,"), but there are ", p, " variable annotations and ", n, " observation annotations.")))
 		end
 
-		# NB: copy because we do not want different DataMatrices to share the same storage
-		var_id_cols = var_id_cols !== nothing ? copy(var_id_cols) : _detect_var_id_cols(var)
-		obs_id_cols = obs_id_cols !== nothing ? copy(obs_id_cols) : _detect_obs_id_cols(obs)
-
-		table_validatecols(var, var_id_cols)
-		table_validatecols(obs, obs_id_cols)
-		table_validateunique(var, var_id_cols)
-		table_validateunique(obs, obs_id_cols)
-		new{T,Tv,To}(matrix, var, obs, var_id_cols, obs_id_cols, models)
+		table_validateunique(var, 1)
+		table_validateunique(obs, 1)
+		new{T,Tv,To}(matrix, var, obs, models)
 	end
 end
 
 
 """
-	DataMatrix(matrix, var, obs; var_id_cols=nothing, obs_id_cols=nothing)
+	DataMatrix(matrix, var, obs)
 
 Create a `DataMatrix` with the given `matrix`, `var` and `obs`.
 
-Columns to use for `var`/`obs` IDs can be explicitly set with `var_id_cols`/`obs_id_cols`.
-Otherwise, an attempt will be made to autodetect the ID columns.
+The first column of `var`/`obs` are used as IDs.
 """
-DataMatrix(matrix, var, obs; var_id_cols=nothing, obs_id_cols=nothing) =
-	DataMatrix(matrix, var, obs, var_id_cols, obs_id_cols, ProjectionModel[])
+DataMatrix(matrix, var, obs) =
+	DataMatrix(matrix, var, obs, ProjectionModel[])
 
 
 """
@@ -106,34 +77,34 @@ function Base.copy(data::DataMatrix; var=:copy, obs=:copy, matrix=:keep)
 	v = var==:copy ? copy(data.var) : data.var
 	o = obs==:copy ? copy(data.obs) : data.obs
 
-	DataMatrix(X, v, o, data.var_id_cols, data.obs_id_cols, copy(data.models))
+	DataMatrix(X, v, o, copy(data.models))
 end
 
 
 
-"""
-	set_var_id_cols!(data::DataMatrix, var_id_cols::Vector{String})
+# """
+# 	set_var_id_cols!(data::DataMatrix, var_id_cols::Vector{String})
 
-Set which column(s) to use as variable IDs.
-The rows of the `data.var` table must be unique, considering only the `var_id_cols` columns.
-"""
-function set_var_id_cols!(data::DataMatrix, var_id_cols::Vector{String})
-	table_validatecols(data.var, var_id_cols)
-	table_validateunique(data.var, var_id_cols)
-	copy!(data.var_id_cols, var_id_cols)
-end
+# Set which column(s) to use as variable IDs.
+# The rows of the `data.var` table must be unique, considering only the `var_id_cols` columns.
+# """
+# function set_var_id_cols!(data::DataMatrix, var_id_cols::Vector{String})
+# 	table_validatecols(data.var, var_id_cols)
+# 	table_validateunique(data.var, var_id_cols)
+# 	copy!(data.var_id_cols, var_id_cols)
+# end
 
-"""
-	set_obs_id_cols!(data::DataMatrix, obs_id_cols::Vector{String})
+# """
+# 	set_obs_id_cols!(data::DataMatrix, obs_id_cols::Vector{String})
 
-Set which column(s) to use as observation IDs.
-The rows of the `data.obs` table must be unique, considering only the `obs_id_cols` columns.
-"""
-function set_obs_id_cols!(data::DataMatrix, obs_id_cols::Vector{String})
-	table_validatecols(data.obs, obs_id_cols)
-	table_validateunique(data.obs, obs_id_cols)
-	copy!(data.obs_id_cols, obs_id_cols)
-end
+# Set which column(s) to use as observation IDs.
+# The rows of the `data.obs` table must be unique, considering only the `obs_id_cols` columns.
+# """
+# function set_obs_id_cols!(data::DataMatrix, obs_id_cols::Vector{String})
+# 	table_validatecols(data.obs, obs_id_cols)
+# 	table_validateunique(data.obs, obs_id_cols)
+# 	copy!(data.obs_id_cols, obs_id_cols)
+# end
 
 
 
@@ -345,9 +316,7 @@ _update_annot(::Any, prefix::String, n::Int) = DataFrame(id=string.(prefix, 1:n)
 """
 	update_matrix(data::DataMatrix, matrix, model=nothing;
 	              var::Union{Symbol,String,DataFrame} = "",
-	              obs::Union{Symbol,String,DataFrame} = "",
-	              var_id_cols,
-	              obs_id_cols)
+	              obs::Union{Symbol,String,DataFrame} = "")
 
 Create a new `DataMatrix` by replacing parts of `data` with new values.
 Mostly useful when implementing new `ProjectionModel`s.
@@ -362,19 +331,15 @@ Kwargs:
   * `::DataFrame` - Replace with a new table with variable annotations.
   * `prefix::String` - Prefix, the new variables will be named prefix1, prefix2, etc.
 * `obs` See `var`.
-* `var_id_cols` - New ID columns. Defaults to the same as data, or "id" if new variables were generated using the "prefix" above.
-* `obs_id_cols` - See `var_id_cols`.
 
 """
 function update_matrix(data::DataMatrix, matrix, model=nothing;
                        var::Union{Symbol,String,DataFrame} = "",
-                       obs::Union{Symbol,String,DataFrame} = "",
-                       var_id_cols = var isa String ? ["id"] : data.var_id_cols,
-                       obs_id_cols = obs isa String ? ["id"] : data.obs_id_cols)
+                       obs::Union{Symbol,String,DataFrame} = "")
 	models = model !== nothing ? vcat(data.models,model) : ProjectionModel[]
 	var = _update_annot(data.var, var, size(matrix,1))
 	obs = _update_annot(data.obs, obs, size(matrix,2))
-	DataMatrix(matrix, var, obs, var_id_cols, obs_id_cols, models)
+	DataMatrix(matrix, var, obs, models)
 end
 
 
@@ -396,30 +361,15 @@ function _printannotation(io, name, show_delim; kwargs...)
 	show_delim && print(io, ", ")
 	printstyled(io, name; kwargs...)
 end
-function _showannotations(io, annotations, id_cols, header)
-	n = names(annotations)
+function _showannotations(io, annotations, header)
 	print(io, header, ": ")
-	actual_id_cols = intersect(n, id_cols)
-	bad_id_cols = setdiff(id_cols, actual_id_cols)
-	other_cols = setdiff(n, actual_id_cols)
-
-	c = 0
-	for name in actual_id_cols
-		_printannotation(io, name, c>0; underline=true)
-		c += 1
-	end
-	for name in bad_id_cols
-		_printannotation(io, name, c>0; underline=true, reverse=true)
-		c += 1
-	end
-	for name in other_cols
+	for (c,name) in enumerate(names(annotations))
 		c>=10 && (print(io, ", ..."); break)
-		_printannotation(io, name, c>0)
-		c += 1
+		_printannotation(io, name, c>1; underline=c==1)
 	end
 end
-_showvar(io, data) = _showannotations(io, data.var, data.var_id_cols, "Variables")
-_showobs(io, data) = _showannotations(io, data.obs, data.obs_id_cols, "Observations")
+_showvar(io, data) = _showannotations(io, data.var, "Variables")
+_showobs(io, data) = _showannotations(io, data.obs, "Observations")
 
 function _showmodels(io, models)
 	print(io, "Models: ")