+++ test/load.jl
@@ -7,25 +7,24 @@
 		@test size(counts)==(P,N)
 		@test nnz(counts.matrix) == expected_nnz
 
-		@test Set(names(counts.obs)) == Set(("id", "barcode"))
-		@test counts.obs.id == expected_barcodes
+		@test Set(names(counts.obs)) == Set(("barcode",))
 		@test counts.obs.barcode == expected_barcodes
 
 		matrix_name = lazy ? "Lazy10xMatrix" : "SparseMatrixCSC"
 		if p==h5_path
 			@test Set(names(counts.var)) == Set(("id", "name", "feature_type", "genome"))
 			@test counts.var.genome == expected_feature_genome
-			test_show(counts; matrix=matrix_name, var=["id", "feature_type", "name", "genome"], obs=["id", "barcode"], models="")
+			test_show(counts; matrix=matrix_name, var=["id", "feature_type", "name", "genome"], obs=["barcode"], models="")
 		else
 			@test Set(names(counts.var)) == Set(("id", "name", "feature_type"))
-			test_show(counts; matrix=matrix_name, var=["id", "feature_type", "name"], obs=["id", "barcode"], models="")
+			test_show(counts; matrix=matrix_name, var=["id", "feature_type", "name"], obs=["barcode"], models="")
 		end
 		@test counts.var.id == expected_feature_ids
 		@test counts.var.name == expected_feature_names
 		@test counts.var.feature_type == expected_feature_types
 
-		@test counts.obs_id_cols == ["id"]
-		@test counts.var_id_cols == ["id", "feature_type"]
+		@test names(counts.obs,1) == ["barcode"]
+		@test names(counts.var,1) == ["id"]
 
 		if lazy
 			@test counts.matrix.filename == p
@@ -59,8 +58,8 @@
 			@test counts.var.name == expected_feature_names
 			@test counts.var.feature_type == expected_feature_types
 
-			@test counts.obs_id_cols == ["id"]
-			@test counts.var_id_cols == ["id", "feature_type"]
+			@test names(counts.obs,1) == ["id"]
+			@test names(counts.var,1) == ["id"]
 
 			if lazy_merge
 				counts = load_counts(counts)
@@ -101,8 +100,8 @@
 			@test counts.var.name == expected_feature_names
 			@test counts.var.feature_type == expected_feature_types
 
-			@test counts.obs_id_cols == ["id"]
-			@test counts.var_id_cols == ["id", "feature_type"]
+			@test names(counts.obs,1) == ["id"]
+			@test names(counts.var,1) == ["id"]
 
 			if lazy_merge
 				counts = load_counts(counts)