+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking
+
+* DataMatrix will now always use the first column of var/obs annotations as ID. (Multiple ID columns are no longer supported.)
+* `load10x` - default to using only first column (id) as unique identifier. Specify e.g. `var_id="var_id"=>["id", "feature_type"]` to merge multiple columns to create the ID.
+* `load10x` - default to using first column (barcode) as unique identifier.
+* `load10x` - no longer supports `copy_obs_col` kwarg.
+
+### Fixed
+
 * Add compat for weakdeps (UMAP, TSne, PrincipalMomentAnalysis).
 
 ## [0.3.9] - 2024-03-04